refactor:change client -> apollo-client
