+++ apps/client/app/employee/[id]/new-review/metric-list.tsx
@@ -2,7 +2,6 @@
 
 import { Button } from '@/components/ui/button'
 import Metric from '@/components/ui/metric'
-
 import { ReportData } from '@/types/models'
 import { useState } from 'react'
 import Link from 'next/link'
